requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.0.4
djangorestframework==3.14.0
joblib==1.1.0
nltk==3.7
Pygments==2.13.0
pytz==2022.4
regex==2022.8.17
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
beautifulsoup4==4.11.1
pandas==1.5.2
geographiclib==2.0
geopy==2.3.0